- update test for controllers  - add test for model

+++ app/controllers/users_controller.rb
@@ -2,13 +2,6 @@ class UsersController < ApplicationController
   # before_action :authenticate_user!, :except => [:follow, :unfollow, :del_tweet]
   protect_from_forgery
 
-  def render_404
-    respond_to do |format|
-      format.html { render :file => "#{Rails.root}/public/404", :layout => false, :status => :not_found }
-      format.xml { head :not_found }
-      format.any { head :not_found }
-    end
-  end
 
   def profile
     @user = User.find_by_fullname(params[:user_name])
@@ -60,8 +53,8 @@ class UsersController < ApplicationController
   def del_tweet
     current_user_id = params[:user_id]
     tweet_id = params[:tweet_id]
-    current_tweet = Tweet.find_by_id(tweet_id)
-    if current_tweet.user_id == current_user_id.to_f
+    if current_user.id == current_user_id.to_i
+      current_tweet = Tweet.find_by_id(tweet_id)
       current_tweet.destroy!
       render json: 'deleted'
     else
@@ -70,29 +63,38 @@ class UsersController < ApplicationController
   end
 
   def follow
-    follow = Follower.new
-    follow.user = User.find_by_id(params[:user])
-    follow.follow = User.find_by_id(params[:follow])
-    if follow.follow.id != follow.user.id
-      if (!Follower.exists?(:user_id => follow.user, :follow_id => follow.follow))
-        follow.save!
-        render json: 'following'
+    if params[:user].to_i == current_user.id
+      follow = Follower.new
+      follow.user = User.find_by_id(params[:user])
+      follow.follow = User.find_by_id(params[:follow])
+      if follow.follow.id != follow.user.id
+        if Follower.exists?(:user_id => follow.user, :follow_id => follow.follow)
+          render json: 'exist'
+        else
+          follow.save!
+          render json: 'following'
+        end
       else
-        render json: 'exist'
+        render json: "You can't be followed for herself`"
       end
     else
-      render json: "You can't be followed for herself`"
+      render json: 'access is denied'
     end
   end
 
   def unfollow
-    user_id = params[:user_id]
-    follow_id = params[:follow_id]
-
-    current_follow = Follower.where(:user_id => user_id, :follow_id => follow_id).first
-    current_follow.destroy!
-
-    render json: 'deleted'
+    if current_user.id == params[:user_id].to_i
+      user_id = params[:user_id]
+      follow_id = params[:follow_id]
+      if (current_follow = Follower.where(:user_id => user_id, :follow_id => follow_id).first).nil?
+        render json: 'error'
+      else
+        current_follow.destroy!
+        render json: 'deleted'
+      end
+    else
+      render json: 'access is denied'
+    end
   end
 
 end
